% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_examples.R
\name{test_amardillo_multithreaded}
\alias{test_amardillo_multithreaded}
\title{runs the multithreaded test from the RcppProgressArmadillo example package}
\usage{
test_amardillo_multithreaded(
  max = 100,
  nb = 1000,
  threads = 0,
  display_progress = TRUE
)
}
\arguments{
\item{max}{the number of loops/increments to execute}

\item{nb}{a parameter controlling the number of computations executed in each loop, so the time 
complexity is quadratic in this parameter}

\item{threads}{the number of OMP threads to use for the computation. If < 0, runs in sequential mode.}

\item{display_progress}{whether to display the progress bar}
}
\value{
the computed number, or -1 if the computation was aborted
}
\description{
runs the multithreaded test from the RcppProgressArmadillo example package
}
