% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_examples.R
\name{test_sequential}
\alias{test_sequential}
\title{run the sequential test from the RcppProgressExample example R package}
\usage{
test_sequential(max = 100, nb = 1000, display_progress = TRUE, ...)
}
\arguments{
\item{max}{the number of loops/increments to execute}

\item{nb}{a parameter controlling the computation executed in each loop}
}
\description{
run the sequential test from the RcppProgressExample example R package
}
