% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_examples.R
\name{test_sequential}
\alias{test_sequential}
\title{runs the sequential test from the RcppProgressExample example R package}
\usage{
test_sequential(max = 100, nb = 1000, display_progress = TRUE)
}
\arguments{
\item{max}{the number of loops/increments to execute}

\item{nb}{a parameter controlling the number of computations executed in each loop, so the time 
complexity is quadratic in this parameter}

\item{display_progress}{whether to display the progress bar}
}
\value{
the computed number, or -1 if the computation was aborted
}
\description{
runs the sequential test from the RcppProgressExample example R package
}
